[refactor/#34] style 패키지 적용
