+++ app/src/main/java/com/smashing/app/core/designsystem/style/DialogStyle.kt
@@ -1,0 +1,6 @@
+package com.smashing.app.core.designsystem.style
+
+enum class DialogStyle {
+    ALERT,
+    CONFIRM;
+}
